Conference Paper Uploads
All conference papers listed in the proposals were found and uploaded to the website. Note that for most of them the link led to the astrophysics data system, which would not allow viewing of the whole article, only the abstract. I'm not totally sure how to access them properly but I've linked the spot where I think the folks with access would be able to login to see the proper thing.